+++ xpilot/src/server/netserver.h
@@ -51,7 +51,7 @@ int Send_score_object(connection_t *connp, double score, int cx, int cy, const c
 int Send_team_score(connection_t *connp, int team, double score);
 int Send_timing(connection_t *connp, int id, int check, int round);
 int Send_base(connection_t *connp, int id, int num);
-int Send_fuel(connection_t *connp, int num, int fuel);
+int Send_fuel(connection_t *connp, int num, double fuel);
 int Send_cannon(connection_t *connp, int num, int dead_time);
 int Send_destruct(connection_t *connp, int count);
 int Send_shutdown(connection_t *connp, int count, int delay);
